[CREATE MEASURES V2 - BACKEND] Step navigation functionality: refactoring and clean up
[TRELLO CARD](https://trello.com/c/IN154EvH/204-dit-create-measures-iteration)
