+++ app/views/workbaskets/create_measures/edit.html.slim
@@ -11,15 +11,15 @@
 
 header
   h1.heading-large
-    = create_measures_header
+    = create_measures_section_header
 
-- if saver_class::FORM_STEPS.include?(params[:step])
+- if step_pointer.current_step_is_form_step?
   = render "workbaskets/create_measures/form"
 - else
   = render "workbaskets/create_measures/steps/review_and_submit"
 
 script
   == "window.save_url = '/create_measures/#{workbasket.id}';"
-  == "window.current_step = '#{params[:step]}';"
+  == "window.current_step = '#{current_step}';"
   == "window.all_settings = '#{workbasket_settings.settings.to_json}';"
   == "window.current_step_settings = '#{workbasket_step_settings.to_json}';"
